django_saml/requirements.txt
asgiref==3.7.2
bleach==6.0.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==41.0.1
defusedxml==0.7.1
Django==3.2
django-cors-headers==4.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djangosaml2==1.6.0
elementpath==4.1.3
gunicorn==20.1.0
idna==3.4
lxml==4.9.2
Markdown==3.3.4
martor==1.6.26
pycparser==2.21
PyJWT==2.7.0
pyOpenSSL==23.2.0
pysaml2==7.4.2
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.3
urllib3==2.0.3
webencodings==0.5.1
xmlschema==2.3.1
